Change exposed ports
Remove 2023, which is insecure telnet CLI and 4569 which is an internal IPC port.

We want to expose the webUI but it is per default on port 8080 and not 80, so
changed that. I've also added a TODO entry about changing the ports in ncs.conf
so we actually expose things on the "correct" ports.