Improve robustness / error handling
`ls *.bin` yields an error when there are no .bin files in the directory and we
avoid this by instead using makes built-in wildcard expansion. For most targets
that rely on the list of NSO install files there will simply be nothing to do if
there are no files. I.e. no hard errors on no .bin files. We do however add a
check that NSO_INSTALL_FILES is not empty in case there is a target that uses it
verbatim and thus relies on it being non-empty.