Add test
This runs a basic test of the built NSO images. We accomplish this by first
taking the development image we've produced and use that to compile the testpkg.
Then we use the production image as a base image, place our testpkg in there to
produce a test docker image (that looks like what a user of this setup would
have). Finally we test this image by running it, just like it would run in a
production setting, for example using a shared volume for persistence data. we
verify the persistence of data by writing some data, stopping and starting a new
NSO container with the same shared volume and then reading the data back.