Dynamically write ncs.conf
We need to set parameters in ncs.conf at runtime and since it is an XML file we
need to modify it in a structured way - not using line oriented tools like sed.
Thus I've incorporated xmlstarlet which provides exactly these facilities for
XML files and used it to set a value. Over time we will modify more things.