Switch to protocol default ports & enable
This switches the northbound interface ports to their protocol defaults, like
NETCONF over SSH to 830 and SSH CLI to 22. The default in the configuration file
are probably to avoid low ports so that NCS can be started without root. We are
however running in a docker container and so we are root. Listening to the
protocol default ports thus makes more sense.

We also enable SSH CLI (22), NETCONF over SSH (830) and NETCONF call-home per
default. The HTTP API is not enabled per default but requires an environment
variable to be passed in. We should have a similar one for HTTPS but it also
requires SSL certificates to be setup so we need to deal with that first.