Now generating ncs.conf at run time
Actually got that in in the last commit but marking TODO entry as done now. I'm
using xmlstarlet to rewrite ncs.conf on startup of the container. xmlstarlet
allows us to deal with the configuration in a structured way, since it is a
structured XML file.

Also using inplace edit together with conditions in the mangle-config shell
script allows us to build a sort of modular way of editing the config controlled
by environment variables.