Revamp test build
Used to spin up the dev container, copy in the source, compile in the
dev container and then copy out the result. That is needlessly
complicated. Now solved using a multi stage docker build instead which
streamlines the process.

I think this should work in CI too, which was one of the reasons for
doing the docker cp in the original solution (rather than a volume).