Use docker registry
The image tags we used meant we only supported local images. Now we support
passing in a DOCKER_REGISTRY variable that sets the registry URL to use. If
CI_REGISTRY_IMAGE is set, which it is in Gitlab CI, then we will use
that (unless DOCKER_REGISTRY is already explicitly set to something).

Fixes #6.