Revamp docker tags
This uses the CI_JOB_ID as the tag on the docker images when building,
like cisco-nso-base:12345 or if you run it locally and CI_JOB_ID isn't
set, it'll use $USER-dev. Tagging on the version like 4.7.5 happens in
the later version-tag make target. That is only done for builds on
master.

We can avoid pushing images by setting DOCKER_PUSH=false.