Change back packages directory to default
I have used /packages for NSO packages instead of the default in a system
install which is /var/opt/ncs/packages. I figured it was cute since we were
running in a container... but it hit me that people might want to be able to
install packages using like Debian packages or similar and for that reason it's
probably easier if we stick to the default installation path even though it's a
little longer.

Fixes #7.