Radically simplify startup script
We've had this status check going in there but it's quite fragile and was
actually broken (killing NCS didn't lead to the container showing as unhealthy
due to a bug). I've now reworked this and made the code much simpler. The
healthcheck has been completely removed for the time being.