Remove unnecessary wait and exit
Since ncs --stop is blocking we don't need to wait for anything... and we dont'
need to exit either as we return to the normal control flow of the program and
that just exits after this.

There is the question about what return code we should return which with this
change is not designed but as I feel there are other ways of ending the
container where we haven't taken an explicit choice on return code, this is good
enough. Also, 143 is for TERM while this handler is also used for INT so we
would need to rework it anyway to be correct. I have issue #20 for figuring out
return status.