Wait for containers to start
It seems that sometimes ncs hasn't been unable to startup by the time we invoke
ncs --wait-started and so we get a connection refused error. I think waiting a
few seconds before should avoid this. Since we are waiting for ncs to start we
are not wasting any time anyway.