Fix the SSH key test
git doesn't track file ownership and privileges so we have to chmod the file for
ssh to accept it (can't be too open).

ssh normally acts based on ~/.ssh/config and since we want to have an isolated
environment for our tests we run ssh with some arguments to this effect.