Add version number to dev tag
Given that the docker image tag we use for local development consists of
the username and -dev, we will overwrite the same tag if we build for
multiple NSO version. Thus, I'm including the NSO version number in the
tag so we instead end up with like kll-dev-4.7.5 or kll-dev-5.2 (for my
username). Much better!