Align wait-started value
We're already doing 600 all over the place. Very unlikely to ever need to go
that high but better to just align and it's nice to have a reasonable default..
let's say if someone else copies our way of running this, then 600 is likely a
good default.

I'm not really any wiser to what the argument to --wait-started does, but I
think specifying a higher value means it will try longer in the initial
connection setup and after that connection has been established but I'm not sure
about the latter part (and the doc says I'm wrong).