Add many "recent" NSO versions
This adds all or at least most of the latest versions of NSO that are still
supported. NCS 3.4 is still supported but I don't think anyone starting to use
NSO in Docker today is interested in such an old version. I'm quite reluctant
about older 4.x releases too but I can't be that arbitrary in what we include
here.