Also generate ed25519 SSH host key
In addition to generating an RSA key, we also generate an ed25519 host
key. NSO 5.3 uses ed25519 per default and will complain over lack of
ed25519. This fixes that and thus we are able to support NSO 5.3.

I'm not sure that it's a wise idea supporting both RSA and ed25519.
Doing an upgrade from something before NSO 5.3 to NSO 5.3 or later will
likely result in a host key change. However, this at least makes it
possible to use NSO 5.3 for the time being.

When was support for ed25519 introduced? Maybe we shold just ditch RSA
completely and use ed25519 instead?