Allow multi-version test by setting OLD_NSO_VERSION
Restructure code around computing the DOCKER_TAG and OLD_DOCKER_TAG a bit. We
first determine our pseudo namespace and then use that to compute the DOCKER_TAG
and OLD_DOCKER_TAG, which is based on OLD_NSO_VERSION, if set.

This means we can run upgrade style testing, i.e. the test cases that support
multi version testing of NSO, by setting OLD_NSO_VERSION and NSO_VERSION to the
two versions we want to test with. To test upgrading from NSO 5.2.1 to 5.3 we
would run:

  make OLD_NSO_VERSION=5.2.1 NSO_VERSION=5.3 test-version