Use Docker image cache for most jobs in CI
So we used to rebuild the entire image, i.e. ignore any cached layers, for all
build jobs running in CI. We want this behaviour because we want to be able to
get all the latest updates, like the latest security patches for our base layer
etc. It's fairly slow though and really only needed for the final images, which
we build from the master branch, thus I'm limiting this so that the cache can
still be used for branches other than master.