Add push stage to CI
The build jobs used to push the final result to a container registry,
which worked and was simple. With the upgrade-test stage we added
asymmetry to the jobs as there could be multiple jobs testing upgrade to
NSO 5.3, or worse, no upgrade test job at all for a particular version
and so trying to push the final result from the upgrade job wouldn't
work. Thus, we are adding a third pipeline stage, called push, which
will push the final image.

Since the push jobes needs to line up with the jobs for the build stage,
we use the same gen.py script to generate include files for the push
jobs. Some internal refatoring was done to the gen.py job to allow for
parameterization of stage and job name as well as running the whole job
config generation as a single function. Rather elegant in the end.

This CI configuration is NOT safe to run across multiple CI runners as
it relies on the Docker images produced by the build stage to be locally
available for the upgrade test and push job stage.