Make multi CI runner compliant
This should make the CI pipeline compatible with running across multiple
CI runners where the different jobs in different stages could be spread
across multiple different runners. For such an environment we have to be
able to pass the docker image built in the first build stage to the jobs
in the other stages. We accomplish that by pulling down the images first
in each job.