Dynamically generate multiver tests
This generates multiversion tests dynamically based on the versions.json
file and some rules. The current rules are a bit of a tryout and
something I came up with in a few minutes thinking what makes sense...
time will tell if this is a good choice. We got quite an combinatorial
effect going here so too permissive rules or too many versions will
result in a really large set of tests.