Ditch 4.4.7
It's just quite old and won't work with --with-package-reload-force, so
it should go.

In theory it would be nice for NSO in Docker to support all currently
supported versions of NSO (which includes 4.3 too!), but I don't think
that's sort of reasonable. Perhaps in the future it will be, but the
current common set of features is too disparate and so we'll just cut
away some of the older versions.