Remove 4.5.9
It just feels rather superfluous and arbitrary to keep this version when
for most other major/minor trains we just have the latest maintenance
release. The exception being 4.7 where we carry both 4.7.5 and 4.7.6 but
I want to show off multi-version testing between maintenance releases in
the same major/minor train and it's the last train for NSO 4 so it is
special.