Force package reloads per default
NSO in Docker is unsupervised so the kind of warnings that a package
reload would emit and that requires the command to be rerun with the
force argument makes little sense. It is better to simply force the
package reload in the first place.

The structured development and operations approach with NSO in Docker
involves testing in CI and operational practices, like backups, to catch
and avoid accidental destructive model changes.