Return correct exit codes upon receiving signals
Been a bit back and forth on this one. In essence, Docker uses Bourne
shell / chroot style exit codes, which means that for child processes
(any PID 1 in a container) that exits due to a signal will get an exit
code of 128+n where n is the signal number, like 143 for SIGTERM (15 on
Linux, and Docker is essentially Linux only).

The Bourne shell does the same thing for its $? variable, which reflects
the exit status of a child process. I originally wanted to do better for
our NSO container since it really is possible, unlike what Bourne does,
to differentiate between the normal exit code 143 and a process
terminating based on SIGTERM. As returned by the waitpid system call, it
is simply completely different bits set for when terminated by signal
and when exiting normally.

ncs has to be stopped gracefully. The ncs binary is actually just the
Erlang BEAM VM renamed and sending a number of signals to it, like
SIGINT, will immediately break the VM. We have to run `ncs --stop` which
will invoke Erlang commands to gracefully shut down ncs. Our container
thus must have signal handlers that catch the signals and run `ncs
--stop`. A process that handles a signal and then returns will exit with
a normal exit code. I figured I could work around this by doing the
needful from the signal handler (i.e. shut down ncs), then uninstalling
the signal handler and finally send the signal to ourselves again which
would then terminate ourselves and we would end up with the "correct"
exit status that reflect we were terminated by a signal. However, this
simply does not work in Docker. I do not know why but as Docker already
masks the whole exit status there's not much point in digging deeper.

We have to work within the environment provided to us by Docker and
so returning 128+n in signal handlers is the best thing we can do. This
commit implements that for what I think is the three most common
signals. Handling more signals is currently out of scope as I don't know
what else to handle.