Mangle username
The tag (DOCKER_TAG) used for docker images or the prefix for container
names (CNT_PREFIX) has a value space that is smaller than what we
currently feed in, which is the username in case of local builds.

This mangles the username so that we only end up with allowed
characters. It doesn theoretically open up for collissions but I feel
that is realistically not going to happen and we can deal with that if
it does.

Anything that isn't permitted is replaced with _, so for example
user@example.com would become user_example.com.