Use custom entrypoint script for dev image
This allows us to easily run NSO related commands as well as get an interactive
shell when no command is specified. I tried playing around with just setting
ENTRYPOINT ["/bin/bash", "-lc"] and combine that with a default CMD but it
doesn't work the way I want it too - the user would have to provide the command
they want to run as a single argument, i.e. they have to quote it. I want to
avoid that, which my special entrypoint script effectively does.