Share commands between -dev and -base
This should achieve a common Docker image layer between the -base and -dev image
for the majority of things we install.

The sacrifice is that we now keep the documentation installed through these
various packages so it seems the -base image increased from 550MB to 557MB for
NSO 5.3. That is acceptable.