Check if set before comparison
Avoid a unary warning message by checking if the CDB_MAJVER variable is
set before comparison. If it's the first time we startup NSO (i.e. no
CDB on disk) then we cannot get the version info from it and thus the
variable will be empty leading to the warning message. This fixes that.