Remove EoM releases
4.7 is the last version that is still maintained. 4.5 and 4.6 are still
supported, which means the support will answer questions on them, but
bug fixes are not backported to them.

Anyone is free to use nso-docker with any NSO version they wish. What
set of versions we use in the public repo is mainly about what versions
of NSO to use when testing public NSO packages on other repositories. I
think it is reasonable that we aim for maintained versions of NSO. Or
seen in a different light, if a FOSS NSO package repo fails CI on NSO
version 4.5, what should be done? If the error is on the NSO side it
won't be fixed as it's not a maintained NSO version. Should the FOSS
package maintainer write backport fixes to ensure the package works with
unmaintained NSO version? That would seem rather unreasonable and not a
sustainable approach for the NSO open source landscape.

Away they go. People should be on 4.7 anyway! ;)