Add a skeleton for NED packages
Similar to the package skeleton, this allows you not only to build
things in an automatic fashion but also start up a test environment or a
development environment. The foundation of this work is based on the
docker-example-ned-ietf repository that I wrote before but it has been
adapted and cleaned up a bit, in particular to have a cleaner structure.

I have also updated the ned-ark to use this skeleton and that has served
as the primary test environment to verify the skeleton itself is working.