Add support for package includes
This makes it possible to include dependencies at build time for the NID package
and NED skeleton. The packages to includes are specified by creating a file in
the includes/ directory. The name of the file will be used as the internal
reference for the package (like inside the Dockerfile) while the content of the
file should point to the Docker image itself. Variables are supported in the
image name.

For example, to include the bgworker package we could create the file
includes/bgworker with the following content:

${INCLUDE_PATH}bgworker:${NSO_VERSION}

At build time it would be expanded into the full path, so if INCLUDE_PATH was
set to 'registry.gitlab.com/nso-developer/' for example, then for a build of NSO
version 5.3 we would grab the package
registry.gitlab.com/nso-developer/bgworker:5.3

For the NID package & NED skeleton, the includes will only be included for the
testnso instance and not in the final output. It is assumed that for a NID
system, includes will end up in the final output. The include packages are not
included in the netsim container image for the NED skeleton, i.e. the includes
can only be used for the test-packages and not for the NED itself. NEDs probably
shouldn't depend on other packages.