Switch to static include path
Using the relative path 'nso-developer/nso-docker' means it will only work for
repositories hosted at gitlab.com, where the nso-docker repo actually is a
gitlab.com/nso-developer/nso-docker. The initial idea was to have the include
path be relative to the local repo, so if someone cloned nso-docker to say
gitlab.com/foo/nso-docker then when building a package repository called mypkg
at gitlab.com/foo/mypkg it would include the version file from
gitlab.com/foo/nso-docker. This isn't possible however as the repo path is a
static string. I tried putting $CI_PROJECT_NAMESPACE (which would have the value
of 'foo' in my example) but there's no variable substitution done on this value.

As it's not possible to get a reference to a local nso-docker repo I am instead
switching to an absolute path. This means we will grab the version include file
from the upstream repo irregardless of where our own project is hosted. This is
an improvement as it will work in more situations and avoid breakage for users.
It is however perhaps a step further away from using a relative link that would
allow the user to easily define their own version include list.