Allow custom Docker registry login
This makes it possible to specify custom credentials to use for authenticating
with the Docker registry through the environment variables CI_DOCKER_USER and
CI_DOCKER_PASSWORD. If the nso-docker repo hosting the base NSO images we need
is a private repo then we need credentials that have access. This is a way of
providing that.

We still use the gitlab-ci-token user for pushing the images since the idea is
that the CI_DOCKER_USER are using a deploy token which only has (read) access on
the nso-docker repo.

Also switched from the deprecated CI_BUILD_TOKEN variable to the newer
CI_JOB_TOKEN.