Align make targets for tag-release
We've used the version-tag make target in the nso-docker repo for putting the
right "release" tag on the docker images. In the skeleton we called it
tag-release. Now we need to align and I think the tag-release and push-release
are better names so I've standardised on that now.

It'll hurt a bit but hopefully not too many users on this yet...