Normalize docker image name and paths
This changes the destination path of the docker images etc that we
output.

We change IMAGE_PATH to point to our project namespace. IMAGE_PATH used
to include the project name when run in CI. As the docker image names
already include the project name, we effectively doubled it in the total
path. That made it harder to write include statements for dependent
projects.

This change makes IMAGE_PATH and PKG_PATH identical. Not sure if that is
good or not. Gives user the chance of including dependencies from one
place and write output somewhere else.