Replace testenv-runcmd with explicit CLI style targets
testenv-runcmd relies on the default style of the NSO instance. Since that could
potentially be changed, we will get non-deterministic behavior. Someone using
testenv-runcmd in a Makefile naturally knows in which CLI style they are writing
their CLI commands. We now make that explicit by calling either testenv-runcmdC
for C-style CLI or testenv-runcmdJ for J-style.

We also take the opportunity to append $(NSO) to the name, which allows us to
pick NSO instance by setting this variable. This will be useful in an LSA
environment where we have multiple NSO instances.