Add mirror mode
A rather strange thing. This allows us to run CI in a special "mirror"
mode. In mirror mode, none of the normal jobs are executed, instead a
special job that acts as a git mirror of this repo from an upstream will
be run. GitLab already has built-in mirror functionality but it
unfortunately requires fast-forward merges to be possible and if you
have done any local modifications to your repository, that's not true.

This mirror script uses git pull and is thus able to mirror a git even
with local changes, although conflict resolution is limited to the
automatic one. Should there be merge conflicts then it will have to be
sorted out manually. The CI job will fail on conflicts and so you should
receive an email about the failed job and can act accordingly.