Only rebuild our own packages
There are other packages in the dst directory, i.e. in the running NSO container
itself, in particular those that were included. We don't need to recompile them
as they were built externally in the first place and there is no guarantee they
even ship with the source code.

We should only build our own packages (including test packages) and that is now
what we do!