Do not run the test suite for the default make target
The default make target is the first target in the Makefile. In our case it is
'all'. Previously, 'all' in turn ran build-all and test-all (actually depended
upon them). Our user instructions are to run 'make' and so this would be what
they saw per default.

The average user that wants to use nso-docker will run it stock, i.e. have no
local modifications and so it is excessive to run the test suite per default.
We can expect that a user that is sufficiently skilled to modify nso-docker also
understands the value of running the test suite.

I added a message to hint the user about the 'test-all' target.

This also has the benefit that the default target now works out of the box on
Mac OS X, as the test suite isn't compatible with the way Docker is run on OS X.