Correct all uses of testenv-runcmd
We recently replaced testenv-runcmd with testenv-runcmdC and testenv-runcmdJ
for explicitly invoking the shell with the given syntax. Now we are updating all
of our uses in skeletons and references in documentation to use the new targets.
Since all examples etc were already using J style, I have converted our uses and
docs to testenv-runcmdJ but testenv-runcmdC is still mentioned in documentation.