Fix skeleton default version set include
Fix the path to the CI config file we include per default. I should've fixed
this with some other config change but it evaded me as my editor didn't grep in
this file after I upgraded to ripgrep... ah the wonders of modern computing.

Anyway, I also improved the help message a bit.