Make docker network creation idempotent
Trying to create a docker network when it already exists leads to an error. We
avoid that big ugly error message which is sure to scare users by turning it
into an idempotent operation. We first check if the network exists with docker
inspect and only if it doesn't do we create the network. We effectively treat
any error from docker inspect as a signal to create the network. There could be
other errors, like not being able to communicate with the docker daemon,
however, in that case we are as likely to see those errors from the invocation
of 'docker network create' instead.