Improve install of prerequisite software
We used to condition the installation of all prerequisite software applications
on the existence of docker. The thinking was that either someone used like
debian:buster as a base, in which case none of the prerequisites, docker least
of all, would be installed. Then our install would be triggered. Alternatively,
the ci-runner-image we build would be used, and it would have all the
prerequisites already installed.

However, in some environment, a CI runner image is used that already has docker
but is missing some of the other prerequisites. The installation procedure is
now slightly rewritten to test for all prerequisites and trigger the install
lacking any of them. A second check is performed for docker, as docker requires
adding a repo and a few other steps. I'm hesitant about its idempotency and
thus wrapped it in this extra check - plus it's probably faster if docker is
already installed too.