Version matching for nightly builds
Relax the rule somewhat, it already matches on such a large portion of the file
anyway that it doesn't matter if we relax this. Now anything can go in as the
version number, as long as it's squeezed in at the right place. This allows
build of nightly builds of NSO as well as other special builds.