Match on ncs- in installer filename too
The way we used to compute the filename of the installer (if the input to us was
just an NSO version) we used to assume the NSO version went in between a few
predefining characters that started with 'nso-'. However, it appears that most
normal NSO releases start with nso- in the filename while nightly builds are
still called ncs-x.y.z.a.. To cater to this, we modify the matching of filenames
so we now also are able to match on ncs- filenames.