Implement robust method of generating Dockerfile
We had problems with our previous solution across GNU and BSD versions of sed. I
made an attempt at using POSIX compliant sed syntax but it really is quite
awkward and I ended up having to write the commands to a file to get newlines
correct (very difficult to write a one-liner with literal newlines in it, which
BSD sed requires after commands like i and a (insert and append respectively)).
This now uses awk instead which is a better fit for this case.