Export path variables
Any variable macros set from within a Makefile is available within the Makefile.
Environment variables are also read in as variables and will be re-exported to
subshells. New variables defined within the Makefile, such as PKG_PATH, are
however not exported per default and so will not be visible to a subshell. The
expansion of the include manifests into the Dockerfile uses a subshell and
requires this be exported.

We could also explicitly set it just for the Dockerfile target but I figure
there might be other targets in the future that might want to access these
values.