Add support for extra files
It's relatively common to want to add in extra files into the container image.
For example, it could be to add things to the configuration by placing a shell
script in /etc/ncs/pre-ncs-start.d/ or something else.

Since using COPY in Docker creates additional layers, we minimize the number of
COPY operations to a single one my having the extra-files directory that is
copied over. Files in the extra-files directory are supposed to contain the
entire path from the root in the Docker image file system. To place 99-foobar.sh
in /etc/ncs/pre-ncs-start.d you would place it in
extra-files/etc/ncs/pre-ncs-start.d/99-foobar.sh