Support cdb-default & bring along aaa_init.xml
This copies over the aaa_init.xml that is installed with NSO from the standard
CDB run directory to a cdb-default directory in the image that is subsequently
loaded in NSO when started with no existing CDB.

Starting the NSO container without any volumes will normally load the
aaa_init.xml file as there is no CDB files and the aaa_init.xml file is in the
CDB directory (it is default NSO behavior to read in the XML files in the CDB
directory). This is the behavior we seek. However, when a shared host volume is
used for the run directory (usually mounting the whole /nso directory), the CDB
directory will be completely empty as the host volume is presumably empty. To
maintain the standard behavior we copy over the aaa_init.xml file into our
/nid/cdb-default directory. On startup we determine if there is a CDB and if
not, all files in /nid/cdb-default will be copied over into the CDB directory
and thus read in by NSO on startup.

This means we now also have the generic ability to load configuration files on
startup if we are starting from scratch. We have a test case to make sure this
works.

The startup script for using default CDB config was inadvertently brought in
with my previous commit. It really belonged with this one. It has now also been
corrected such that it works properly.